Create 'Area and funding' integration test
This leaves the "view tasklist" feature much smaller,
focussed only on the presentation of tasks within sections.

We will go on to add a description of new section and
its initial task:

section: "About Firstname Lastname"
task: "Complete equality & diversity monitoring"

This commit also re-words some of the commentary to make the
language more consisted with BDD terms:

- Given: the initial set up state
- When: the action whose results are under consideration
- Then/And: the results which we observe
